Add de.lucaswerkmeister.ctsl.source_gen module
Parses the ECMAScript and TypeScript grammars, which are assumed to be
present in the root directory of the repository. They may be downloaded
from the URLs given in the assertion messages that are displayed if the
files can’t be found.

(Generating a lexer and parser isn’t done yet.)