Add removeLeftRecursion function stub
It doesn’t actually do anything, but the documentation tells you how the
conversion would work.

---

However, this is probably where this iteration of CTSL ends. As outlined
in https://github.com/ceylon/ceylon/issues/3041#issuecomment-176239905,
we’ll go for another approach. All the commits up to this point will be
stowed away on some branch and perhaps eventually deleted.