Add initial TypeScript project
ctsl.ts demonstrates how to use TypeScript compiler or Node.js APIs. The
Makefile can be used to run the program (which is a bit tricky, because
TypeScript doesn’t have its own module runtime, but the compiler isn’t
written to be used in common.js).

The symbolic links for DefinitelyTyped and TypeScript are necessary
because the TypeScript compiler doesn’t accept .. in reference paths.