Convert simple .d.ts file to .ceylon file
This converts the fairly simple definition file for acc-wizard to
Ceylon. The conversion is done at the AST level; an alternative
approach, working on the model instead, didn’t work out so well, but
some WIP code is stowed away in ctsl-model.ts just in case it’s needed
later.

Of course, we originally wanted to emit a model file… the ceylon file
won’t stay forever. But first things first – now I know how to work on
the TypeScript compiler AST. Emitting the model is next.