Add support for type parameters
No constraints yet, and if you try to actually use the type parameters,
it won’t work, because we currently assume that every non-native type is
in the current package. But at least this already works at runtime,
because the reified type arguments are appended to the regular
arguments, not prepended as on the JVM backend, so they’re just ignored.