Add support for type arguments and Array<> type
This is where it all falls apart. In the JS model, type arguments are
named: they are addressed by the type parameter name, not by index. But
without typechecking the AST, we don’t know the type parameter name.

The solution, for now, is to assume that the type parameter name always
equals the type argument name, and then to add some special cases for
some types, like Array (where we know the type parameter is called
Element). I hope I’ll be able to make this work for the first iteration
of the program (which only has to load the TypeScript compiler’s model).
(The second iteration will, for this and other reasons, operate on a
typechecked model.)

Also adds a special case that Array is from the language module.