Add test for index signatures
As it turns out, I can’t add runtime support for this in CTSL: adding
something to Object.prototype (here: the $_get function) confuses the
runtime in various ways that I don’t understand. (Here, module loading
seems to break, so that ceylon.test can’t launch the test module; adding
it later seems to break reified type checks.)

Instead, we add the function to the prototype right before we need it,
and remove it immediately after.

(Like a lot of other stuff in here, this is a horrible hack that’s only
for the first version.)