Emit runtime type information
For interfaces and type aliases, we now emit JS code identical to
ceylon-js’ output for equivalent Ceylon types.

For enums, we have to simplify a bit, because at runtime they’re still
just integers, so they can’t have the right prototype or stuff.

This is necessary to assign a TypeScript value to a TypeScript type from
a dynamic block, i. e. to satisfy dre$$’ checks.